+++ assets/js/devs/fitvids.js
@@ -14,7 +14,7 @@ var FitVids = (function () {
 	var Constructor = function (selectors) {
 
 		// Default video players
-		var defaults = 'iframe[src*="player.vimeo.com"], iframe[src*="youtube.com"]';
+		var defaults = 'iframe[src*="player.vimeo.com"], iframe[src*="youtube.com"], iframe[src*="youtube-nocookie.com"], iframe[src*="kickstarter.com"][src*="video.html"], object, embed';
 
 		// Public APIs object
 		var publicAPIs = {};